changed routes back to original and moved the belongsto into the index.js inside models

+++ routes/api/product-routes.js
@@ -1,10 +1,10 @@
-const router = require('express').Router();
-const { Product, Category, Tag, ProductTag } = require('../../models');
+const router = require("express").Router();
+const { Product, Category, Tag, ProductTag } = require("../../models");
 
 // The `/api/products` endpoint
 
 // get all products
-router.get('/', (req, res) => {
+router.get("/", (req, res) => {
   // find all products
   Product.findAll({
     include: [{ model: Category }, { model: Tag }],
@@ -14,23 +14,23 @@ router.get('/', (req, res) => {
 });
 
 // get one product
-router.get('/:id', (req, res) => {
+router.get("/:id", (req, res) => {
   // find a single product by its `id`
   Product.findByPk(req.params.id, {
-    include: [{ model: Category }, {model: Tag}],
+    include: [{ model: Category }, { model: Tag }],
   })
-  .then((product) => {
-    if (!product) {
-      res.status(404).json({ message: 'No product with this id!' });
+    .then((product) => {
+      if (!product) {
+        res.status(404).json({ message: "No product with this id!" });
         return;
-    }
-    res.json(product);
-  })
-  .catch((err) => res.status(500).json(err));
+      }
+      res.json(product);
+    })
+    .catch((err) => res.status(500).json(err));
 });
 
-// create new product
-router.post('/', (req, res) => {
+// creat new product
+router.post("/", (req, res) => {
   /* req.body should look like this...
     {
       product_name: "Basketball",
@@ -42,7 +42,7 @@ router.post('/', (req, res) => {
   Product.create(req.body)
     .then((product) => {
       // if there's product tags, we need to create pairings to bulk create in the ProductTag model
-      if (req.body.tagIds.length) {
+      if (req.body.tagIds && req.body.tagIds.length) {
         const productTagIdArr = req.body.tagIds.map((tag_id) => {
           return {
             product_id: product.id,
@@ -62,7 +62,7 @@ router.post('/', (req, res) => {
 });
 
 // update product
-router.put('/:id', (req, res) => {
+router.put("/:id", (req, res) => {
   // update product data
   Product.update(req.body, {
     where: {
@@ -70,49 +70,50 @@ router.put('/:id', (req, res) => {
     },
   })
     .then((product) => {
-      // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
+      if (req.body.tagIds && req.body.tagIds.length) {
+        const productTags = ProductTag.findAll({
+          where: { product_id: req.params.id },
         });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
+        const productTagIds = productTags.map(({ tag_id }) => tag_id);
+        // create filtered list of new tag_ids
+        const newProductTags = req.body.tagIds
+          .filter((tag_id) => !productTagIds.includes(tag_id))
+          .map((tag_id) => {
+            return {
+              product_id: req.params.id,
+              tag_id,
+            };
+          });
+        // figure out which ones to remove
+        const productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+          .map(({ id }) => id);
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+        // run both actions
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove } }),
+          ProductTag.bulkCreate(newProductTags),
+        ]);
+      }
+
+      return res.json(product);
     })
-    .then((updatedProductTags) => res.json(updatedProductTags))
     .catch((err) => {
       // console.log(err);
       res.status(400).json(err);
     });
 });
 
-router.delete('/:id', (req, res) => {
+router.delete("/:id", (req, res) => {
   Product.destroy({
     where: { id: req.params.id },
   })
     .then((product) => {
       if (!product) {
-        res.status(404).json({ message: 'No product with this id!' });
+        res.status(404).json({ message: "No product with this id!" });
         return;
       }
-      res.json({ message: 'Successfully deleted product!!' });
+      res.json({ message: "Successfully deleted product!!" });
     })
     .catch((err) => res.status(500).json(err));
 });
